+++ trunk/src/y-thr/tests/test-team.cpp
@@ -70,7 +70,7 @@ YOCTO_UNIT_TEST_IMPL(team)
 	size_t ncycle = 8;
 	size_t length = 800000;
 	if( argc > 1 ) ncycle = atol( argv[1] );
-	if( argc > 2 ) length = strtod( argv[2], NULL );
+	if( argc > 2 ) length = size_t(strtod( argv[2], NULL ));
 	std::cerr << ">>>> team ncycle=" << ncycle << ", length=" << length << std::endl;
